Add a new recovery operation, partial reconstruction
 Partial reconstruction is a method that recovers when a crash occurs between a Version Update operation and an Aggregation Transaction.
This method reads the same page from the original database file and writes it back to the shadow database file.
Specifically, read the GPL from the header page of the shadow database file, read the GPL from the header page of the original database file, and merge it.
Using the result of the merged contents, read pages from the original database file to memory and write it back to the shadow database file.
Finally, the version number of the header page in the shadow database file is reset to zero, indicating that the partial reconstruction operation has been completed.

Signed-off-by: sundoo Kim <sksioi12@hanyang.ac.kr>